etl/sqlite_to_postgres/requirements.txt
psycopg2-binary==2.9.9
Faker==20.1.0
python-dotenv==1.0.0
pytest==7.4.3